+++ RELEASENOTES.md
@@ -140,6 +140,7 @@
         ([#2646]()https://github.com/androidx/media/issues/2646).
     *   Allow codec reuse for EAC3, EAC3-JOC and AC-4 formats
         ([#1346](https://github.com/androidx/media/pull/1346)).
+    *   Add support for float PCM samples in `Sonic`.
 *   Video:
 *   Text:
 *   Metadata: